Use native type checking; remove checkType()
Also removed undocumented (I think) behavior of returning the current canvas when target() is called without a param. Since the method can accept either a canvas or its context, presumably the consumer would expect target() to return whatever it was called with, so removing this behavior altogether to avoid confusion.